services/pdf-parser/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
pdfminer.six==20251107
